backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1 
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1